api/requirements.txt
Flask==3.1.2
langchain_community==0.3.29
langchain_core==0.3.75
langchain_ollama==0.3.7
psycopg2-binary==2.9.10
python-dotenv==1.1.1
SQLAlchemy==2.0.43
Gunicorn==22.0.0
